processor-service/requirements.txt
fastapi[standard]
dlt[s3, postgres]>=1.7.0
boto3==1.36.23
python-decouple
pandas
openpyxl
streamlit